完成后台内容分类的增删查改，以及商城首页的页面显示

+++ e3-manager/e3-manager-service/src/main/java/cn/e3mall/service/impl/ItemServiceImpl.java
@@ -52,7 +52,7 @@ public class ItemServiceImpl implements ItemService {
 		
 		//获得总数
 		List<TbItem> itemList = info.getList();
-		int size = (int) info.getTotal();
+		long size = info.getTotal();
 		//封装结果集
 		DataGridResult result = new DataGridResult();
 		result.setRows(itemList);
